started new challenges page, add new content data
