+++ app/page.tsx
@@ -1,6 +1,5 @@
 import Hero from "./sections/hero";
 import Stakeholders from "./sections/stakeholders";
-import Challenges from "./sections/challenges";
 import Team from "./sections/team";
 import Awards from "./sections/awards";
 import Contact from "./sections/contact";
@@ -12,7 +11,6 @@ export default function Home() {
       <main>
         <Hero />
         <Stakeholders/>
-        <Challenges />
         <Team />
         <Awards />
         <Contact />
